Tracing contract improvements

+++ contracts/tracer/src/msg.rs
@@ -12,9 +12,13 @@ pub enum ExecuteMsg {
     Forward {
         addr: String,
         msg: Binary,
+        #[serde(default)]
         marker: u64,
+        #[serde(default)]
         catch_success: bool,
+        #[serde(default)]
         catch_failure: bool,
+        #[serde(default)]
         fail_reply: bool,
     },
     Clear {},
@@ -24,7 +28,7 @@ pub enum ExecuteMsg {
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
 pub enum QueryMsg {
-    Log { depth: Option<u64> },
+    Log { depth: Option<u32> },
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
